backend/requirements.txt
flask
sqlmodel
chromadb
google-generativeai
python-dotenv
pytest
flask-sqlalchemy
flask-login
gunicorn
boto3
psycopg2-binary
yt-dlp
